Markdown UIを別コンポーネントに切り出し
